+++ packages/node-core/CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix handle bigint type in jsonb array, both get and set method will generate consistent result (#2495)
-
+- Update with `@subql/utils`, fixed json and json array hashcode issue, improve poi consistency (#2495)
 ### Added
 - SubqueryProject base from extracing common code (#2496)
+
 
 ## [12.0.0] - 2024-07-22
 ### Changed